+++ packages/@runlightyear/slack/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @runlightyear/slack
 
+## 0.14.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @runlightyear/lightyear@0.34.0
+
+## 0.14.3
+
+### Patch Changes
+
+- Updated dependencies [7469fbe82]
+  - @runlightyear/lightyear@0.33.0
+
+## 0.14.2
+
+### Patch Changes
+
+- Updated dependencies [7e667cda1]
+  - @runlightyear/lightyear@0.32.0
+
 ## 0.14.1
 
 ### Patch Changes